+++ tabletest-junit/CHANGELOG.md
@@ -1,13 +1,12 @@
 # TableTest-Junit Changelog
 
 ## [Unreleased]
+
+## [0.2.0] - 2025-05-22
 ### Added
 - Support for the first column being a scenario name
 - Support for set values in tables
 - Support for expanding set values to one test per value
-
-## [0.2.0]
-### Added
 - Support for loading table from external file
 
 ## [0.1.0] - 2025-05-11